fix: affine maps not updated

+++ hesaffnet/LAF.py
@@ -101,7 +101,7 @@ def checkTouchBoundary(LAFs):
     if LAFs.is_cuda:
         pts = pts.cuda()
     out_pts =  torch.bmm(LAFs_to_H_frames(LAFs),pts.expand(LAFs.size(0),3,4))[:,:2,:]
-    good_points = 1 -(((out_pts > 1.0) +  (out_pts < 0.0)).sum(dim=1).sum(dim=1) > 0)
+    good_points = 1 -(((out_pts > 1.0) +  (out_pts < 0.0)).sum(dim=1).sum(dim=1) > 0).type(torch.float)
     return good_points
 
 def bsvd2x2(As):
